+++ nodes/EasyRedmine/fields/PersonalContactFields.ts
@@ -129,8 +129,7 @@ export const PersonalContactFields: INodeProperties[] = [
 		displayName: 'EasyRedmine Personal Contacts Query Name or ID',
 		name: 'personalContactQueryId',
 		type: 'options',
-		description:
-			'Choose a query to filter the results. Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code-examples/expressions/">expression</a>.',
+		description: 'Choose a query to filter the results. Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code-examples/expressions/">expression</a>. Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>.',
 		displayOptions: {
 			show: {
 				resource: [EasyNodeResourceType.personalContacts],